- also remove instance - move tests to separate folder - eslint endOfLine: auto
