+++ .eslintrc.json
@@ -38,7 +38,7 @@
     "no-prototype-builtins": "off",
     "no-restricted-globals": "off",
     "no-underscore-dangle": "off",
-    "prettier/prettier": "error",
+    "prettier/prettier": ["error", { "endOfLine": "auto" }],
     "semi": "off",
     "standard/no-callback-literal": "off"
   }
